requirements.txt
Flask==2.1.0
psycopg2-binary==2.9.1
Flask-CORS==3.1.1